% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_outcome.R
\name{predict_outcome}
\alias{predict_outcome}
\title{A function that predicts something}
\usage{
predict_outcome(x)
}
\arguments{
\item{x}{Uniformity Cell Size}
}
\value{
Outcome of tumor prediction (malignant/benign)
}
\description{
A function that predicts something
}
